pyproject.toml
[project]
name = "newsletter-automation"
version = "0.1.0"
description = "AI-powered newsletter automation for Automata using LangGraph deepagents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.1.0",
    "tavily-python>=0.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
